Unconditionally suffix the CHECKSUM.SHA256 and CHECKSUM.MD5 files with the OSRELEASE-TARGET_ARCH.  When copying the BETA and RC builds to ftp-master, the CHECKSUM.{SHA256,MD5} files are clobbered with the new version, which prevents us from having more than one -BETA or -RC build available on FTP at any given time.
Sponsored by:	The FreeBSD Foundation


git-svn-id: file:///Users/gjb/Downloads/releng-svn/thermite-r688@235 5711f056-dcf5-4cc1-a278-612e92bcda7e

+++ user/gjb/scripts/release/ftp-stage.sh
@@ -175,6 +175,10 @@ stage_isos() {
 			echo "${newname}.img.bz2"
 			mv ${oldname}.img.bz2 ${newname}.img.bz2
 		fi
+	fi
+
+	if [ "X${newname}" != "X${oldname}" ]; then
+		cd ${C}/R
 		rm -f CHECKSUM.SHA256* CHECKSUM.MD5*
 		# CHECKSUM.SHA256-11.0-CURRENT-amd64-VT-20140127-r261200
 		echo "=== Regenerating SHA256 checksums"
@@ -185,6 +189,12 @@ stage_isos() {
 		md5 ${__DISCNAME}* > \
 			${C}/R/CHECKSUM.MD5-${_sumsuffix}
 		cd ${scriptdir}
+	else
+		for h in SHA256 MD5; do
+			echo "=== Renaming ${h} checksums"
+			mv ${C}/R/CHECKSUM.${h} \
+				${C}/R/CHECKSUM.${h}-${_sumsuffix}
+		done
 	fi
 
 	# Copy ISO images to FTP snapshots directory.
